+++ package/example/lib/advanced_example/main.dart
@@ -5,6 +5,7 @@ import 'package:easy_form_example/advanced_example/screens/login_demo/login_demo
 import 'package:easy_form_example/advanced_example/screens/save_indicator/save_indicator_screen.dart';
 import 'package:easy_form_example/advanced_example/screens/save_error/save_error_screen.dart';
 import 'package:easy_form_example/advanced_example/screens/clear_value/clear_value_screen.dart';
+import 'package:easy_form_example/advanced_example/screens/error_displayer/error_displayer_screen.dart';
 
 void main() {
   runApp(MyApp());
@@ -68,6 +69,11 @@ class StartScreen extends StatelessWidget {
                   _DemoButton(
                     text: 'Clear value',
                     builder: (context) => ClearValueScreen(),
+                  ),
+                  const SizedBox(height: 16),
+                  _DemoButton(
+                    text: 'Error displayer',
+                    builder: (context) => ErrorDisplayerScreen(),
                   ),
                 ],
               ),